+++ Cargo.toml
@@ -6,9 +6,12 @@ edition = "2024"
 [dependencies]
 cargo_metadata = "0.20.0"
 clap = { version = "4.0", features = ["derive"] }
-serde = {version = "1.0.219",features = ["derive"] }
+dircpy = "0.3.19"
+serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.99"
 toml = "0.8.23"
+
+[target.'cfg(target_os = "windows")'.dependencies]
 windows = { version = "0.61.3", features = [
     "Win32_Foundation",
     "Win32_UI_Shell",
@@ -18,3 +21,6 @@ windows = { version = "0.61.3", features = [
     "Win32_System",
     "Win32_System_Threading",
 ] }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+nix = { version = "0.30", features = ["user"] }
