+++ src/main.rs
@@ -1,16 +1,17 @@
 mod command;
-use crate::command::{MV, Install};
+mod mv;
+
 use crate::command::{Cargo, JKCommand};
 use cargo_metadata::Message;
 use clap::Parser;
 use serde::{Deserialize, Serialize};
 use serde_json;
-use std::{env, fs};
+use std::env;
 use std::io;
 use std::io::Write;
-use std::path::Path;
 use std::process::Command;
 use std::process::Stdio;
+
 #[derive(Debug, Deserialize)]
 struct CargoToml {
     package: Package,
@@ -41,10 +42,11 @@ fn main() {
     let Cargo::Input(input) = Cargo::parse();
     // let ostype = env::consts::OS;
     // println!("Operating System: {}", ostype);
-    let aesdk_root =
-        env::var("AESDK_ROOT").expect("AESDK_ROOT is not defined as an environment variable");
     match input.cmd {
         JKCommand::Build(build) => {
+            let _aesdk_root = env::var("AESDK_ROOT")
+                .expect("AESDK_ROOT is not defined as an environment variable");
+
             // load Cargo.toml and read the metadata
             let current_dir = env::current_dir().expect("Failed to get current directory");
             let cargo_toml_path = current_dir.join("Cargo.toml");
@@ -115,8 +117,8 @@ fn main() {
             }
         }
         JKCommand::MV(mv) => {
-            if platform::is_elevated() {
-                match mv_command(&mv) {
+            if mv::is_elevated() {
+                match mv::mv_command(&mv) {
                     Ok(_) => eprintln!("File moved successfully."),
                     Err(e) => eprintln!("Failed to move file: {e}"),
                 }
@@ -126,49 +128,32 @@ fn main() {
                 let _ = io::stdin().read_line(&mut String::new());
             } else {
                 eprintln!("Not running with elevated privileges. Attempting to elevate...");
-                platform::elevate_self();
+                mv::elevate_self();
             }
-        },
+        }
         JKCommand::Install(_install) => {
             install_command();
-        },
+        }
     }
-}
-
-fn mv_command(mv: &MV) -> io::Result<()> {
-    let target_dir = Path::new("C:\\Program Files\\Adobe\\Common\\Plug-ins\\7.0\\MediaCore\\");
-
-    let src = Path::new(&mv.src);
-    // get filename from the source path
-    let filename = src.file_name().ok_or(io::Error::new(
-        io::ErrorKind::NotFound,
-        "Source file does not have a valid filename",
-    ))?;
-    let target = target_dir.join(filename);
-
-    fs::copy(src, target)
-        .map_err(|e| io::Error::other(format!("Failed to move file: {e}")))?;
-
-    Ok(())
 }
 
 fn install_command() {
     eprintln!("Starting install process...");
-    
+
     // Detect if we're running in development mode (cargo run -- jk) or production mode (cargo jk)
     let current_exe = env::current_exe().expect("Failed to get current executable path");
     let is_dev_mode = current_exe.to_string_lossy().contains("target");
-    
+
     let (cmd_prefix, cmd_args): (&str, Vec<&str>) = if is_dev_mode {
         ("cargo", vec!["run", "--", "jk"])
     } else {
         ("cargo", vec!["jk"])
     };
-    
+
     // Step 1: Execute build command
     let build_cmd = format!("{} {} build --format json", cmd_prefix, cmd_args.join(" "));
     eprintln!("Running: {}", build_cmd);
-    
+
     let mut build_command = Command::new(cmd_prefix);
     for arg in &cmd_args {
         build_command.arg(arg);
@@ -178,43 +163,43 @@ fn install_command() {
         .arg("--format")
         .arg("json")
         .output();
-    
+
     match build_output {
         Ok(output) => {
             if !output.status.success() {
                 eprintln!("Build failed: {}", String::from_utf8_lossy(&output.stderr));
                 std::process::exit(1);
             }
-            
+
             // Step 2: Parse the JSON output to get the aex file path
             let stdout = String::from_utf8_lossy(&output.stdout);
             eprintln!("Build output: {}", stdout);
-            
+
             match serde_json::from_str::<AexFileOutput>(&stdout) {
                 Ok(aex_output) => {
                     let aex_file = &aex_output.aex_file;
                     eprintln!("Built aex file: {}", aex_file);
-                    
+
                     // Step 3: Execute mv command
                     let mv_cmd = format!("{} {} mv {}", cmd_prefix, cmd_args.join(" "), aex_file);
                     eprintln!("Running: {}", mv_cmd);
-                    
+
                     let mut mv_command = Command::new(cmd_prefix);
                     for arg in &cmd_args {
                         mv_command.arg(arg);
                     }
-                    let mv_output = mv_command
-                        .arg("mv")
-                        .arg(aex_file)
-                        .output();
-                    
+                    let mv_output = mv_command.arg("mv").arg(aex_file).output();
+
                     match mv_output {
                         Ok(mv_result) => {
                             if mv_result.status.success() {
                                 eprintln!("Install completed successfully!");
                                 eprintln!("{}", String::from_utf8_lossy(&mv_result.stdout));
                             } else {
-                                eprintln!("Move failed: {}", String::from_utf8_lossy(&mv_result.stderr));
+                                eprintln!(
+                                    "Move failed: {}",
+                                    String::from_utf8_lossy(&mv_result.stderr)
+                                );
                                 std::process::exit(1);
                             }
                         }
@@ -236,84 +221,4 @@ fn install_command() {
             std::process::exit(1);
         }
     }
-}
-
-#[cfg(target_os = "windows")]
-mod platform {
-    use std::ffi::OsStr;
-    use std::iter::once;
-    use std::os::windows::ffi::OsStrExt;
-    use windows::{
-        core::*,
-        Win32::{Foundation::*, Security::*, System::Threading::*, UI::{Shell::*, WindowsAndMessaging::SW_SHOW}},
-    };
-
-    pub fn is_elevated() -> bool {
-        let mut token = HANDLE::default();
-        unsafe {
-            if OpenProcessToken(GetCurrentProcess(), TOKEN_QUERY, &mut token).is_ok() {
-                let mut elevation = TOKEN_ELEVATION::default();
-                let mut size = std::mem::size_of::<TOKEN_ELEVATION>() as u32;
-                if GetTokenInformation(
-                    token,
-                    TokenElevation,
-                    Some(&mut elevation as *mut _ as *mut core::ffi::c_void),
-                    size,
-                    &mut size,
-                ).is_ok() {
-                    CloseHandle(token).unwrap();
-                    return elevation.TokenIsElevated != 0;
-                }
-            }
-        }
-        false
-    }
-
-pub fn elevate_self() {
-    let exe_path = std::env::current_exe().unwrap();
-
-    let args: Vec<String> = std::env::args().skip(1).collect(); // 最初は自分自身なので skip
-    let arg_str = args
-        .into_iter()
-        .map(|arg| {
-            if arg.contains(' ') {
-                format!("\"{arg}\"")
-            } else {
-                arg
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(" ");
-
-    let cmd = OsStr::new(exe_path.to_str().unwrap())
-        .encode_wide()
-        .chain(once(0))
-        .collect::<Vec<u16>>();
-
-    let params = OsStr::new(&arg_str)
-        .encode_wide()
-        .chain(once(0))
-        .collect::<Vec<u16>>();
-
-    let mut sei = SHELLEXECUTEINFOW {
-        cbSize: std::mem::size_of::<SHELLEXECUTEINFOW>() as u32,
-        lpVerb: PCWSTR(w!("runas").as_ptr()),
-        lpFile: PCWSTR(cmd.as_ptr()),
-        lpParameters: PCWSTR(params.as_ptr()),
-        nShow: SW_SHOW.0,
-        fMask: SEE_MASK_NOCLOSEPROCESS,
-        ..Default::default()
-    };
-
-    unsafe {
-        if ShellExecuteExW(&mut sei).is_ok() {
-            WaitForSingleObject(sei.hProcess, INFINITE);
-            CloseHandle(sei.hProcess).unwrap();
-        } else {
-            eprintln!("Failed to elevate via UAC");
-        }
-    }
-
-    std::process::exit(0);
-}
 }
